pyproject.toml
[tool.poetry]
name = "llm-prompt-recurse"
version = "0.1.0"
description = ""
authors = ["Erik Nomitch <enomitch@gmail.com>"]
license = "GPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-anthropic = "^0.1.4"
langchain = {extras = ["all"], version = "^0.1.13"}
click = "^8.1.7"
python-dotenv = "^1.0.1"
ipython = "^8.22.2"
langchainhub = "^0.1.15"
langsmith = "^0.1.31"
ipdb = "^0.13.13"
yaspin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
